Split out HelloTriangle
