+++ apps/benchmarking-test-app/src/graphicsBenchmarks/HelloTriangle.ts
@@ -1,0 +1,100 @@
+// https://webgpu.github.io/webgpu-samples/?sample=helloTriangle
+import { redFragWGSL, triangleVertWGSL } from "./triangle";
+import { CanvasContext } from "./types";
+
+export const runHelloTriangle = async (
+  context: CanvasContext,
+  device: GPUDevice,
+  requestAnimationFrame: (callback: (time: number) => void) => number
+) => {
+  const presentationFormat = navigator.gpu.getPreferredCanvasFormat();
+  const pipeline = device.createRenderPipeline({
+    layout: "auto",
+    vertex: {
+      module: device.createShaderModule({
+        code: triangleVertWGSL,
+      }),
+      entryPoint: "main",
+    },
+    fragment: {
+      module: device.createShaderModule({
+        code: redFragWGSL,
+      }),
+      entryPoint: "main",
+      targets: [
+        {
+          format: presentationFormat,
+        },
+      ],
+    },
+    primitive: {
+      topology: "triangle-list",
+    },
+  });
+
+  const uniformBufferSize = 4;
+  const uniformBuffer = device.createBuffer({
+    size: uniformBufferSize,
+    usage: GPUBufferUsage.UNIFORM | GPUBufferUsage.COPY_DST,
+  });
+
+  const uniformBindGroup = device.createBindGroup({
+    layout: pipeline.getBindGroupLayout(0),
+    entries: [
+      {
+        binding: 0,
+        resource: {
+          buffer: uniformBuffer,
+        },
+      },
+    ],
+  });
+
+  return new Promise<void>((resolve) => {
+    let frame = 0;
+    const animate = () => {
+      const textureView = context.getCurrentTexture().createView();
+
+      const renderPassDescriptor: GPURenderPassDescriptor = {
+        colorAttachments: [
+          {
+            view: textureView,
+            clearValue: [0, 0, 0, 1],
+            loadOp: "clear",
+            storeOp: "store",
+          },
+        ],
+      };
+      const commandEncoder = device.createCommandEncoder();
+
+      const floatBuffer = new Float32Array(1);
+      floatBuffer.fill(frame, 0);
+      frame++;
+
+      device.queue.writeBuffer(
+        uniformBuffer,
+        0,
+        floatBuffer.buffer,
+        floatBuffer.byteOffset,
+        floatBuffer.byteLength
+      );
+
+      const passEncoder = commandEncoder.beginRenderPass(renderPassDescriptor);
+      passEncoder.setPipeline(pipeline);
+      passEncoder.setBindGroup(0, uniformBindGroup);
+      passEncoder.draw(3);
+      passEncoder.end();
+
+      device.queue.submit([commandEncoder.finish()]);
+      context.present();
+
+      if (frame >= 500) {
+        resolve();
+      } else {
+        requestAnimationFrame(() => animate());
+      }
+    };
+
+    requestAnimationFrame(animate);
+  });
+};
