--- apps/benchmarking-test-app/src/graphicsBenchmarks/TriangleWGPUBenchmark.tsx
@@ -1,136 +1,0 @@
-import { StyleSheet, View, PixelRatio } from "react-native";
-import { Canvas, useCanvasEffect } from "react-native-wgpu";
-
-import { redFragWGSL, triangleVertWGSL } from "./triangle";
-
-export const TriangleWGPUBenchmark: React.FC<{
-  onComplete: (startTime: number) => void;
-}> = ({ onComplete }) => {
-  const ref = useCanvasEffect(async () => {
-    const startTime = Date.now();
-    const adapter = await navigator.gpu.requestAdapter();
-    if (!adapter) {
-      throw new Error("No adapter");
-    }
-    const device = await adapter.requestDevice();
-    const presentationFormat = navigator.gpu.getPreferredCanvasFormat();
-
-    const context = ref.current!.getContext("webgpu")!;
-    const canvas = context.canvas as HTMLCanvasElement;
-    canvas.width = canvas.clientWidth * PixelRatio.get();
-    canvas.height = canvas.clientHeight * PixelRatio.get();
-
-    if (!context) {
-      throw new Error("No context");
-    }
-
-    context.configure({
-      device,
-      format: presentationFormat,
-      alphaMode: "opaque",
-    });
-
-    const pipeline = device.createRenderPipeline({
-      layout: "auto",
-      vertex: {
-        module: device.createShaderModule({
-          code: triangleVertWGSL,
-        }),
-        entryPoint: "main",
-      },
-      fragment: {
-        module: device.createShaderModule({
-          code: redFragWGSL,
-        }),
-        entryPoint: "main",
-        targets: [
-          {
-            format: presentationFormat,
-          },
-        ],
-      },
-      primitive: {
-        topology: "triangle-list",
-      },
-    });
-
-    const uniformBufferSize = 4;
-    const uniformBuffer = device.createBuffer({
-      size: uniformBufferSize,
-      usage: GPUBufferUsage.UNIFORM | GPUBufferUsage.COPY_DST,
-    });
-
-    const uniformBindGroup = device.createBindGroup({
-      layout: pipeline.getBindGroupLayout(0),
-      entries: [
-        {
-          binding: 0,
-          resource: {
-            buffer: uniformBuffer,
-          },
-        },
-      ],
-    });
-
-    let frame = 0;
-    const animate = () => {
-      const textureView = context.getCurrentTexture().createView();
-
-      const renderPassDescriptor: GPURenderPassDescriptor = {
-        colorAttachments: [
-          {
-            view: textureView,
-            clearValue: [0, 0, 0, 1],
-            loadOp: "clear",
-            storeOp: "store",
-          },
-        ],
-      };
-      const commandEncoder = device.createCommandEncoder();
-
-      const floatBuffer = new Float32Array(1);
-      floatBuffer.fill(frame, 0);
-      frame++;
-
-      device.queue.writeBuffer(
-        uniformBuffer,
-        0,
-        floatBuffer.buffer,
-        floatBuffer.byteOffset,
-        floatBuffer.byteLength
-      );
-
-      const passEncoder = commandEncoder.beginRenderPass(renderPassDescriptor);
-      passEncoder.setPipeline(pipeline);
-      passEncoder.setBindGroup(0, uniformBindGroup);
-      passEncoder.draw(3);
-      passEncoder.end();
-
-      device.queue.submit([commandEncoder.finish()]);
-      context.present();
-
-      if (frame >= 500) {
-        onComplete(startTime);
-      } else {
-        requestAnimationFrame(() => animate());
-      }
-    };
-
-    requestAnimationFrame(animate);
-  });
-
-  return (
-    <View style={style.container}>
-      <Canvas ref={ref} style={style.webgpu} />
-    </View>
-  );
-};
-
-const style = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  webgpu: {
-    flex: 1,
-  },
-});
