+++ apps/benchmarking-test-app/src/graphicsBenchmarks/types.ts
@@ -1,0 +1,3 @@
+export type CanvasContext = GPUCanvasContext & {
+  present: () => void;
+};
